+++ samples/shared/src/commonMain/kotlin/eric/bitria/samples/App.kt
@@ -8,7 +8,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import eric.bitria.samples.scenes.AnimationScene
 import eric.bitria.samples.scenes.DuckScene
-import eric.bitria.samples.scenes.EnvironmentScene
+import eric.bitria.samples.scenes.HDREnvironmentScene
+import eric.bitria.samples.scenes.KTXEnvironmentScene
 import eric.bitria.samples.scenes.PickingScene
 import eric.bitria.samples.scenes.PrimitivesScene
 import eric.bitria.samples.scenes.SolarScene
@@ -26,7 +27,8 @@ fun App() {
             Screen.Solar      -> SolarScene(onBack = { screen = Screen.Home })
             Screen.Animation  -> AnimationScene(onBack = { screen = Screen.Home })
             Screen.SplitView  -> SplitViewScene(onBack = { screen = Screen.Home })
-            Screen.Environment -> EnvironmentScene(onBack = { screen = Screen.Home })
+            Screen.KTXEnvironment -> KTXEnvironmentScene(onBack = { screen = Screen.Home })
+            Screen.HDREnvironment -> HDREnvironmentScene(onBack = { screen = Screen.Home })
         }
     }
 }