+++ samples/shared/src/commonMain/kotlin/eric/bitria/samples/Navigation.kt
@@ -9,5 +9,6 @@ sealed class Screen {
     data object Solar : Screen()
     data object Animation : Screen()
     data object SplitView : Screen()
-    data object Environment : Screen()
+    data object KTXEnvironment : Screen()
+    data object HDREnvironment : Screen()
 }